requirements.txt
tensorflow==2.2.0
matplotlib==3.1.3
keras==2.1.5
Pillow==7.0.0
numpy==1.19.0
pandas==1.0.5
scikit-image==0.16.2
scipy==1.2.0
opencv-python==4.2.0

